대댓글이 있을 때 모댓글 삭제 불가

+++ src/main/java/com/example/CUSHProject/service/CommentService.java
@@ -25,6 +25,7 @@ public class CommentService {
 
     /*댓글 등록*/
     public void commentPost(BoardCommentDto boardCommentDto, String username) throws Exception {
+
         Optional<MemberEntity> memberEntity = memberRepository.findByUsername(username);
         Optional<BoardEntity> boardEntity = boardRepository.findById(boardCommentDto.getBoardId());
 
@@ -32,6 +33,7 @@ public class CommentService {
         boardCommentDto.setUpdateDate(LocalDateTime.now());
 
         BoardCommentEntity boardCommentEntity = boardCommentDto.toEntity();
+
 
         boardCommentEntity.setCommentBoardId(boardEntity.get());
         boardCommentEntity.setCommentWriter(memberEntity.get());
@@ -42,7 +44,8 @@ public class CommentService {
     /*댓글 리스트*/
     public List<BoardCommentDto> getCommentList(Long bid) throws Exception {
         Optional<BoardEntity> boardEntity = boardRepository.findById(bid);
-        List<BoardCommentEntity> boardCommentEntityList = boardCommentRepository.findByBoardId(boardEntity.get());
+        //List<BoardCommentEntity> boardCommentEntityList = boardCommentRepository.findByBoardId(boardEntity.get());
+        List<BoardCommentEntity> boardCommentEntityList = boardCommentQueryRepository.findByBoardId(boardEntity.get());
         List<BoardCommentDto> boardCommentDtoList = new ArrayList<>();
 
         for(int i = 0; i < boardCommentEntityList.size(); i++) {
@@ -52,11 +55,37 @@ public class CommentService {
     }
 
     /*댓글 삭제*/
-    public void deleteComment(Long cid){
-        boardCommentRepository.deleteById(cid);
+    public Long deleteComment(Long cid) {
+        Optional<BoardCommentEntity> boardCommentEntity = boardCommentRepository.findById(cid);
+        int depth = boardCommentEntity.get().getCDepth();
+        Long reCommentCnt = boardCommentQueryRepository.deleteCheck(cid);
+
+        /*if (boardCommentEntity.get().getWriter().getRole().getValue().equals("ROLE_ADMIN")) {
+            boardCommentRepository.deleteById(cid);
+        } else*/
+        if (depth == 0) {
+            if (boardCommentQueryRepository.deleteCheck(cid) == 0) {
+                boardCommentRepository.deleteById(cid);
+            }
+        } else if (depth == 1) {
+            boardCommentRepository.deleteById(cid);
+        }
+
+        return reCommentCnt;
     }
 
     public void modifyComment(Long cid, String comment){
         boardCommentQueryRepository.updateComment(cid,comment);
+    }
+
+    /*대댓글 목록*/
+    public List<BoardCommentDto> getReCommentList(Long cid) {
+        List<BoardCommentEntity> boardCommentEntityList = boardCommentQueryRepository.findByCGroup(cid);
+        List<BoardCommentDto> boardCommentDtoList = new ArrayList<>();
+
+        for(int i = 0; i < boardCommentEntityList.size(); i++) {
+            boardCommentDtoList.add(boardCommentEntityList.get(i).toDto());
+        }
+        return boardCommentDtoList;
     }
 }
